added cast

+++ interpolate.h
@@ -34,7 +34,7 @@ public:
   template <bool fold=false>
   struct Lut {
     // maybe want a.size() since we truncate
-    Lut(const Vector& a) : A(a), lgScale(std::max(0L, lg(A.size() - 1UL) + lg((uint64_t)A.back()) - 64L)) {
+    Lut(const Vector& a) : A(a), lgScale(std::max(0L, lg((uint64_t)A.size() - 1UL) + lg((uint64_t)A.back()) - 64L)) {
       if (fold) divisors /= (A.size() - 1);
       d_range_width = DivLut::Gen(A.back() - A[0]) / (A.size() - 1);
     }
